完成登录注册部分功能，更改数据库设计脚本

+++ tm-repo/src/main/java/com/zwp/travelmemories/repo/mybatis/mappers/UserMapper.java
@@ -26,7 +26,7 @@ public interface UserMapper {
      * @param vo
      * @return
      */
-    Long insertUser(UserVo vo);
+    int insertUser(UserVo vo);
